fix  ui

+++ components/Data-table.tsx
@@ -36,8 +36,10 @@ import {
   Columns3Icon,
   EllipsisIcon,
   ListFilterIcon,
+  Loader,
   PinOffIcon,
   PlusIcon,
+  Trash,
   TrashIcon,
 } from "lucide-react";
 
@@ -84,6 +86,7 @@ import { Iheads } from "@/app/dashboard/[workspaceId]/form/view/[formId]/_compon
 import { mutate } from "swr";
 import { toast } from "sonner";
 import { apiClient } from "@/lib/axios";
+import { useParams } from "next/navigation";
 
 type Item = {
   id: string;
@@ -129,21 +132,27 @@ export default function TanStackTable({
   };
   formId: string;
 }) {
-  const id = useId();
   const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
   const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({});
   const inputRef = useRef<HTMLInputElement>(null);
 
-
   const [data, setData] = useState<any[]>(tableData);
 
-  const handleDeleteRows = () => {
-    const selectedRows = table.getSelectedRowModel().rows;
-    const updatedData = data.filter(
-      (item) => !selectedRows.some((row) => row.original?.id === item?.id)
-    );
-    setData(updatedData);
-    table.resetRowSelection();
+  const handleDeleteRows = async () => {
+    try {
+      const selectedRows = table.getSelectedRowModel().rows;
+      const updatedData = data.filter(
+        (item) => !selectedRows.some((row) => row.original?.id === item?.id)
+      );
+      const selectedRowsId = selectedRows.map((r) => r?.original?.id);
+      await apiClient.put(`/api/respondent/multiple`, selectedRowsId);
+      mutate(`/api/response/form/${formId}`);
+      setData(updatedData);
+      table.resetRowSelection();
+      toast("selected rows deleted !");
+    } catch (e) {
+      toast("failed to delete selected rows");
+    }
   };
 
   const table = useReactTable({
@@ -646,16 +655,22 @@ export default function TanStackTable({
 }
 
 export function RowActions({ row }: { row: Row<any> }) {
+  const [deleting, setDeleting] = useState(false);
+
+  const { formId } = useParams();
   const handleDelete = async (id: string) => {
-    console.log(id);
+    setDeleting(true);
 
     if (!id) return;
     try {
       await apiClient.delete(`/api/respondent/${id}`);
+      mutate(`/api/response/form/${formId}`);
       toast("respondent deleted !");
     } catch (e) {
       toast("failed to delete");
     }
+
+    setDeleting(false);
   };
 
   return (
@@ -673,20 +688,38 @@ export function RowActions({ row }: { row: Row<any> }) {
         </div>
       </DropdownMenuTrigger>
       <DropdownMenuContent align="end">
-        {/* <DropdownMenuGroup>
-          <DropdownMenuItem>
-            <span>Edit</span>
-            <DropdownMenuShortcut>⌘E</DropdownMenuShortcut>
-          </DropdownMenuItem>
-        </DropdownMenuGroup> */}
-        <DropdownMenuSeparator />
-        <DropdownMenuItem
-          onClick={() => handleDelete(row?.original?.id)}
-          className="text-destructive focus:text-destructive"
-        >
-          <span>Delete</span>
-          <DropdownMenuShortcut>⌘⌫</DropdownMenuShortcut>
-        </DropdownMenuItem>
+        <AlertDialog>
+          <AlertDialogTrigger className="hover:bg-accent focus:text-accent-foreground data-[variant=destructive]:text-destructive data-[variant=destructive]:focus:bg-destructive/10 dark:data-[variant=destructive]:focus:bg-destructive/20 data-[variant=destructive]:focus:text-destructive data-[variant=destructive]:*:[svg]:!text-destructive [&_svg:not([class*='text-'])]:text-muted-foreground relative cursor-default rounded-sm px-2 py-1.5 text-sm outline-hidden select-none data-[disabled]:pointer-events-none data-[disabled]:opacity-50 data-[inset]:pl-8 [&_svg]:pointer-events-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4 w-full flex items-center gap-2">
+            <Trash />
+            <p>Delete</p>
+          </AlertDialogTrigger>
+          <AlertDialogContent>
+            <div className="flex flex-col gap-2 max-sm:items-center sm:flex-row sm:gap-4">
+              <div
+                className="flex size-9 shrink-0 items-center justify-center rounded-full border"
+                aria-hidden="true"
+              >
+                <CircleAlertIcon className="opacity-80" size={16} />
+              </div>
+              <AlertDialogHeader>
+                <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+                <AlertDialogDescription>
+                  This action cannot be undone. This will permanently delete
+                  response
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+            </div>
+            <AlertDialogFooter>
+              <AlertDialogCancel>Cancel</AlertDialogCancel>
+              <AlertDialogAction
+                onClick={() => handleDelete(row?.original?.id)}
+              >
+                {deleting && <Loader className=" animate-spin" />}
+                Delete
+              </AlertDialogAction>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialog>
       </DropdownMenuContent>
     </DropdownMenu>
   );
